Fix Errors: Connvert Double to Float (X.XX) to (X.XXf)
Looking for some Speed and memory upgrades
-Connvert Int(4 Byte) to uint16_t(2 Byte)
(For Mobs per Stage uint8_t because only 1-9)

Update Project Properties:
-Full spped and small code (full speed high prio)

Database Upgrade:
-Items Storage in struct Data List
-Not fully load form file (to-do-list)
-ENUMs for special Types of Items (Rare,Class, ...)

If i have time i will write it to end.

+++ Lord of the dark Tavern/dungeon_manager.cpp
@@ -77,12 +77,12 @@ void dungeon_manager::fill_dungeon(const std::vector<sf::Image*> &image)
 		boost::random::mt19937 dev{ static_cast<std::uint32_t>(times) };;
 		mobs_ebene = dev()*(i+1) % 9 + 1;
 
-		for (int u = mobs_ebene; u != 0; u--)
+		for (uint16_t u = mobs_ebene; u != 0; u--)
 		{
 			std::time_t fro = std::time(0);
 			boost::random::mt19937 dev_test{ static_cast<std::uint32_t>(fro) };;
 			//std::cout <<"Vector:" << image.capacity() << std::endl;
-			int rnd = dev_test()*u % (image.capacity()-2);
+			uint16_t rnd = dev_test()*u % (image.capacity()-2);
 			std::cout << rnd << std::endl;
 			mob_ebene_1.push_back(new mob(u-1,*image[rnd], diffculty, 0, pfont, *image[image.capacity()-1]));
 		}
